Replace square/go-jose with go-jose/go-jose
The https://github.com/square/go-jose repository was archived on
February 27, 2023.

`ory/fosite` uses `go-jose/go-jose` as a replacement in accordance with
the `square/go-jose` maintainer's recommendation. Since we rely on
`ory/fosite` in order to implement distrust, I've decided that
the same version of `go-jose/go-jose` that `ory/fosite` uses would be an
appropriate replacement as well.

This also lets us avoid the GO-2024-2631 vulnerability that was
reported in v2.6.0.